Revert "포스트 모달 관련 PR입니다. (#10)"
This reverts commit fedc516687a70e42082bd24c08f6cdd236417557.
