+++ src/components/Modal/PostModal.jsx
@@ -1,89 +1,103 @@
 import { useState } from 'react';
 import styled from 'styled-components';
 import { useAuth } from '../../contexts/Auth/hooks';
-import { useModal } from '../../contexts/Modal/useModal';
+import { useDispatch } from 'react-redux';
+import { closeModal } from '../../Redux/Slices/PostModalSlice';
+import { addPosts } from '../../Redux/Slices/PostsSlice';
+import { v4 as uuidv4 } from 'uuid';
 
-export default function PostModal({ data }) {
-  const { id: prevId, title: prevTitle, content: prevContent, img_url: prevImg_url, created_at: prevCreated_at } = data;
+export default function PostModal() {
+  const [postTitle, setPostTitle] = useState('');
+  const [postContent, setPostContent] = useState('');
+  const [postImgFile, setPostImgFile] = useState(null);
+  const [previewUrl, setPreviewUrl] = useState('');
 
-  const { supabaseClient } = useAuth();
+  const dispatch = useDispatch();
+  const { session, supabaseClient } = useAuth();
 
-  const [editedData, setEditedData] = useState({
-    id: prevId,
-    title: prevTitle,
-    content: prevContent,
-    img_url: prevImg_url,
-    created_at: prevCreated_at
-  });
-  const [isEditMode, setIsEditMode] = useState(false);
+  const handleImageChange = (event) => {
+    const fileObj = event.target.files[0];
+    setPostImgFile(fileObj);
+    const objectUrl = URL.createObjectURL(fileObj);
+    setPreviewUrl(objectUrl);
+  };
 
-  const [_, closeModal] = useModal();
+  async function handlePost(e) {
+    e.preventDefault();
+    const { data: imgData, error: imgError } = await supabaseClient.storage
+      .from('posts')
+      .upload(`thumbnail/${Date.now()}`, postImgFile);
 
-  function handleEnableEditMode() {
-    setIsEditMode(true);
-  }
-
-  async function handleEditPost() {
-    if (confirm('정말 게시물을 수정하시겠어요?')) {
-      const { error } = await supabaseClient
-        .from('posts')
-        .update({ ...editedData })
-        .eq('id', prevId);
-      if (error) {
-        throw new Error(error);
-      }
-      alert('수정이 완료되었습니다.');
-    } else {
-      setEditedData(data);
+    if (imgError) {
+      throw new Error(imgError);
     }
-    setIsEditMode(false);
-  }
 
-  async function handleDeletePost() {
-    if (confirm('정말 게시물을 삭제하시겠어요?')) {
-      await supabaseClient.from('posts').delete().eq('id', prevId);
-      alert('삭제가 완료되었습니다.');
-      closeModal();
-    }
+    const {
+      data: { publicUrl }
+    } = supabaseClient.storage.from('posts').getPublicUrl(imgData.path);
+
+    const newPost = {
+      id: uuidv4(),
+      UID: session.user.user_metadata.userName,
+      title: postTitle,
+      content: postContent,
+      img_url: publicUrl
+    };
+    // posts DB에 삽입
+    await supabaseClient.from('posts').insert(newPost);
+    dispatch(addPosts(newPost));
+
+    alert('포스팅 완료!');
+    dispatch(closeModal());
   }
 
   return (
     <StModalWrapper>
-      <img src={prevImg_url} width={200} height={200} />
-      {isEditMode ? (
-        <>
-          <input
-            type="text"
-            value={editedData.title}
-            onChange={(e) => setEditedData({ ...editedData, title: e.target.value })}
-          />
-          <input
-            type="text"
-            value={editedData.content}
-            onChange={(e) => setEditedData({ ...editedData, content: e.target.value })}
-          />
-        </>
-      ) : (
-        <>
-          <p>{editedData.title}</p>
-          <p>{editedData.content}</p>
-        </>
-      )}
-      <button
-        onClick={isEditMode ? handleEditPost : handleEnableEditMode}
-        disabled={isEditMode && prevTitle === editedData.title && prevContent === editedData.content}
-      >
-        수정하기
-      </button>
-      <button onClick={handleDeletePost}>삭제하기</button>
-      <button onClick={closeModal}>모달 닫기</button>
+      <StFormWrapper onSubmit={handlePost}>
+        <label htmlFor="postImage">
+          <StPrevImgWrapper>
+            {previewUrl ? (
+              <img src={previewUrl} alt="미리보기 이미지" width={360} height={180} />
+            ) : (
+              <p>이미지 가져오기</p>
+            )}
+          </StPrevImgWrapper>
+        </label>
+        <StFileInput type="file" id="postImage" accept="image/*" onChange={handleImageChange} />
+        <input
+          type="text"
+          value={postTitle}
+          placeholder="제목을 입력해주세요."
+          onChange={(e) => setPostTitle(e.target.value)}
+        />
+        <input
+          type="text"
+          value={postContent}
+          placeholder="간략한 소개글을 작성해주세요."
+          onChange={(e) => setPostContent(e.target.value)}
+        />
+        <button type="submit">포스팅하기</button>
+        <button onClick={() => dispatch(closeModal())}>모달창 닫기</button>
+      </StFormWrapper>
     </StModalWrapper>
   );
 }
 
 const StModalWrapper = styled.div`
-  width: 600px;
-  height: 400px;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  z-index: 1000;
+  width: 600px; /* 예시 크기 */
+  height: 400px; /* 예시 크기 */
+  padding: 20px;
+  background-color: white;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  border-radius: 10px;
+`;
+
+const StFormWrapper = styled.form`
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -92,4 +106,23 @@ const StModalWrapper = styled.div`
   padding: 32px;
   border-radius: 10px;
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+`;
+
+const StPrevImgWrapper = styled.div`
+  width: 400px;
+  height: 200px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.3);
+  cursor: pointer;
+  border-radius: 8px;
+  transition: all 0.2s ease-in-out;
+  &:hover {
+    opacity: 0.7;
+  }
+`;
+
+const StFileInput = styled.input`
+  display: none;
 `;
